BTSS tests work with persistent allocator

+++ src/tools/include/memoria/tools/tests.hpp
@@ -44,8 +44,7 @@ class TestTask: public Task {
 protected:
     typedef Task Base;
 
-    Int     size_;
-
+    BigInt	size_;
     String  current_test_name_;
 
     RngInt  	int_generator_;
